+++ routes/categories.js
@@ -47,6 +47,8 @@ router.put("/:id", [auth, admin, validateObjectId], async (req, res) => {
 
 // INFO: delete category route
 router.delete("/:id", [auth, admin, validateObjectId], async (req, res) => {
+  if (req.user.role !== "super") return res.status(401).send("Access denied.");
+
   const category = await Category.findByIdAndRemove(req.params.id);
 
   if (!category)
